Change OSC 'b' (Blob) encoding to parse, and produce strings of bytes.
Previously 'b' would only encode single numbers 0..255, and decode
Blob fields as lists of numbers 0..255, collapsed into the list of
arguments, so a non-terminal 'b' value could not be determined
easily.  The C* pack required multiple arguments to correctly pack
a stream of bytes greater than 1, replacing 'C' with 'a' in the b
pack/unpack option allows parse and message to work with bytestrings.